feat: mantain the navbar position fixed
