+++ web/lib/zk_arcade_web/components/layouts/app.html.heex
@@ -1,6 +1,6 @@
 <main class="min-h-screen flex flex-col">
-  <div class="section-width w-full" style="max-width: 1000px">
-    <div class="absolute left-0 w-full mt-10" style="z-index: 1">
+  <div class="nav-container fixed pt-10 w-full flex justify-center z-[1000] transition-all duration-300">
+    <div class="w-full" style="max-width: 1000px">
       <.nav 
         network={Application.get_env(:zk_arcade, :network)}
         payment_service_address={Application.get_env(:zk_arcade, :payment_service_address)}
@@ -10,7 +10,7 @@
         username={@conn.assigns.username}
         user_position={@conn.assigns.user_position}
         explorer_url={@explorer_url}
-       />
+      />
     </div>
   </div>
